call-translator/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.2.0
attrs==23.2.0
bidict==0.22.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.3
deepgram-sdk==3.1.1
deepl==1.16.1
Flask==3.0.1
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
gTTS==2.5.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
marshmallow==3.20.2
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
PyAudio==0.2.14
pydub==0.25.1
python-dotenv==0.14.0
python-engineio==4.8.2
python-socketio==5.11.0
requests==2.31.0
setuptools==69.0.3
simple-websocket==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.25
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.2.0
verboselogs==1.7
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.4
